fix ruff docstring warnings
